+++ DataService/Dtos/Queries/MetricValueQuery.cs
@@ -4,7 +4,7 @@ public sealed record MetricValueQuery
 {
     public Ulid Id { get; init; } = Ulid.Empty;
     public string? Tag { get; init; }
-    public int? UserId { get; init; }
+    public long? UserId { get; init; }
     public Guid? UserGroupId { get; init; }
     public string? MetricName { get; init; }
     public DateTimeOffset? StartDate { get; init; }
